Adding relog
